#21 feat: implement the user login UI and function
When the user click on the button user in the first page or the main page of our app it takes the user to the login page where can the user put in its cardinalities in the right field and then click on the login button if the login been verified it provide the user to access  other part of the app. we have tow kind of user General and admin.

Co-authored-By: Elif Yaqoub Gussayel@student.gu.se