#1 Refactor: Improving the code with conditionals to control lights if statements to control the lights based on flag variable instead of using repetitive delays and timer counts
This commit improves the code by use of conditional if statements to control the lights based on flag variable instead of using repetitive delays and timer count variables and having to increment each one of them. This improves the readability and maintainability of the code. It also makes it more DRY and more in real time lights are switched.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se