#1 Fix: Fix bug with traffic light code not running after a cycle of 40flags is complete
This commit fixes the bug by resetting the flag variable to 0 after it reaches 40. Before this the code was running for turning traffic lights on and off for the first 40 seconds as the delay is set to 1000ms which increments the flag variable by one every 1000ms. But after the 40 seconds it won't work before due to the conditional statements which are bounded within 0-40 variable value. For that reason to make the traffic lights continue working smoothly without interruption, flag variable will be reset after every flag=40.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se