#1 feat: Added Serial Display Prints for Traffic Light Control System
This commit adds serial display prints to the normal traffic light system. On the display, time-lapse is shown for the traffic lights so that the user can know for how long they have to wait. In this refactored version rather than having the time lapse for only the red light, this feature has a more real-time approach and is expected to be more useful as it shows the timings for all lights: Red, yellow and green. The same pattern of 20sec for Red, 3 for yellow and 17 for green is followed.This feature will also help with monitoring the traffic light system.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se