#10 feat: Added a buzzer to make a loud sound
This commit sets up a buzzer as output and then puts it in the loop so that the buzzer can ring almost every second. Once it rings with frequency 1000 then 2000 and then stops and the cycle repeats itself. This is implemented so that the user can find their car by hearing the buzzer sound.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se