#10 feat: Added a button for buzzer
This commit creates up a button for buzzer. And this button is on topå of the car. The colour was not really going well with it so now we have the it invisible over the top of the car in car console page. When user will click the car it will turn on the car finding button functionality.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se