#10 feat: Made the button be used for both turning the buzzer on AND off
This commit adds conditionals for checking in and changing the state of the button on or off and then implementing the function accordingly.  Added clicklistener to handle button presses. And it is the basic attempt to connect the button with this functionality and trying to maintain a connection with mqtt within this car console page.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se