#10 fix: Fixed bugs in arduinocode
This commit fixes the bugs that were not letting the code compile and execute properly for car finding in arduino. There were problems related to parenthesis and cases of variable names used were inconsistent which have now been fixed along with replacing some functions in setup rather than in loop.

Co-Authored-By: Tehreem Asif gusasite@student.gu.se