#27 feat: Implement splash screen
implemented the lottie animation view in splash screen class This code is using the (postDelayed()) method of the (Handler) class in Java to delay the execution of the code inside the (run()) method by 5000 milliseconds (or 5 seconds). In this case, the code inside the (run()) method is creating a new (Intent) to start the (MainActivity) of the application and then starting the activity using the (startActivity()) method.
 Create and design the layout by adding the element in it .
Co-Authored-By: Elif Yaqoub  gussayel@student.gu.se